+++ examples/src/env.d.ts
@@ -1,5 +1,4 @@
-declare module '*.png';
-
+/// <reference types="ts-lib-tools" />
 interface Pointer {
   x: number;
   y: number;