+++ packages/ts-lib-scripts/assets/common-template/src/env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="ts-lib-tools" />